move paywall-builder and stories inside src/
