+++ paywall/.storybook/config.js
@@ -12,7 +12,7 @@ setConfig({
   },
 })
 
-const req = require.context('../stories', true, /\.stories\.js$/)
+const req = require.context('../src/stories', true, /\.stories\.js$/)
 
 function loadStories() {
   req.keys().forEach(filename => req(filename))
