+++ paywall/rollup.paywall.config.js
@@ -3,7 +3,7 @@ import babel from 'rollup-plugin-babel'
 import { uglify } from 'rollup-plugin-uglify'
 
 const config = {
-  input: 'paywall-builder/index.js',
+  input: 'src/paywall-builder/index.js',
   output: {
     format: 'umd',
     name: 'paywall.min.js',
