+++ paywall/src/paywall-builder/blocker.js
