+++ paywall/src/paywall-builder/mutationObserver.js
