+++ paywall/src/paywall-builder/script.js
